Update end_date forms for new simpler designs
