+++ templates/barriers/edit/end_date.html
@@ -1,0 +1,35 @@
+{% extends "barriers/edit/base.html" %}
+
+{% block page_title %}{{ block.super }} - Edit end date{% endblock %}
+
+{% block back_link %}
+    <a href="{% url 'barriers:barrier_detail' barrier.id %}" class="govuk-back-link ma-back-link">Back</a>
+{% endblock %}
+
+{% block page_content %}
+
+    {% include 'partials/heading.html' with text='Update the Barrier' %}
+    {% form_error_banner form %}
+
+    <form action="" method="POST">
+        {% csrf_token %}
+
+        <div id="{{ form.end_date.name }}" class="govuk-form-group{% if form.end_date.errors %} govuk-form-group--error{% endif %}">
+            <fieldset class="govuk-fieldset" {% if form.end_date.errors %}aria-desribedby="status-date-group-error"{% endif %} role="group">
+                <legend class="govuk-fieldset__legend govuk-fieldset__legend--s">{{ form.end_date.label }}</legend>
+
+                <span class="govuk-hint">{{ form.end_date.help_text }}</span>
+
+                {% form_field_error form "end_date" %}
+
+                <div class="govuk-date-input" id="status-date-group">
+                    {{ form.end_date }}
+                </div>
+            </fieldset>
+        </div>
+
+        <input type="submit" class="govuk-button" value="Save and exit to barrier">
+        <a class="form-cancel" href="{% url 'barriers:barrier_detail' barrier.id %}">Cancel</a>
+    </form>
+
+{% endblock %}
