--- templates/barriers/edit/partials/end_date.html
@@ -1,34 +1,0 @@
-<div id="{{ form.has_end_date.name }}" class="govuk-form-group{% if form.has_end_date.errors %} govuk-form-group--error{% endif %}">
-    <fieldset class="govuk-fieldset">
-        <legend class="govuk-fieldset__legend govuk-fieldset__legend--s">{{ form.has_end_date.label }}</legend>
-
-        {% form_field_error form "has_end_date" %}
-
-        <div class="govuk-radios {{ form.has_end_date.name }} govuk-radios--conditional" data-module="radios">
-            {% for value, name in form.fields.has_end_date.choices %}
-                <div class="govuk-radios__item">
-                    <input class="govuk-radios__input" id="{{ form.has_end_date.name }}-{{ forloop.counter }}" name="{{ form.has_end_date.name }}" type="radio" value="{{ value }}"{% if form.has_end_date.value == value %} checked="checked"{% endif %}>
-                    <label class="govuk-label govuk-radios__label" for="{{ form.has_end_date.name }}-{{ forloop.counter }}">{{ name }}</label>
-                </div>
-
-                {% if value == "yes" %}
-                    <div class="govuk-radios__conditional govuk-radios__conditional--hidden" id="conditional-{{ form.has_end_date.name }}-{{ value }}">
-                        <div id="{{ form.end_date.name }}" class="govuk-form-group{% if form.end_date.errors %} govuk-form-group--error{% endif %}">
-                            <fieldset class="govuk-fieldset" {% if form.end_date.errors %}aria-desribedby="status-date-group-error"{% endif %} role="group">
-                                <legend class="govuk-fieldset__legend govuk-fieldset__legend--s">{{ form.end_date.label }}</legend>
-
-                                <span class="govuk-hint">{{ form.end_date.help_text }}</span>
-
-                                {% form_field_error form "end_date" %}
-
-                                <div class="govuk-date-input" id="p">
-                                    {{ form.end_date }}
-                                </div>
-                            </fieldset>
-                        </div>
-                    </div>
-                {% endif %}
-            {% endfor %}
-        </div>
-    </fieldset>
-</div>
